# perf(marshal): Memoize checkRemotable