Without this, marshaling a remotable incurs a redundant verification in `getInterfaceOf`.

<details><summary>ad hoc analysis</summary>

```sh
node --input-type=module -e '
  import Benchmark from "benchmark";
  import "../init/index.js";
  import { makeMarshal } from "./src/marshal.js";
  const { serialize } = makeMarshal(x => x, x => x, { errorTagging: "off", serializeBodyFormat: "smallcaps" });
  let result;
  (new Benchmark.Suite())
    .add("shallow remotable", {
      setup: () => {
	const rProto = Object.defineProperties({}, { [Symbol.for("passStyle")]: { value: "remotable" }, [Symbol.toStringTag]: { value: "Remotable" } });
        shallow = Object.create(rProto);
        Array(20).fill().forEach((_, i) => { shallow[`method${i}`] = () => {}; });
      },
      fn: () => { result = serialize(harden(shallow)); },
    })
    .add("deep remotable", {
      setup: () => {
	const rProto = Object.defineProperties({}, { [Symbol.for("passStyle")]: { value: "remotable" }, [Symbol.toStringTag]: { value: "Remotable" } });
        deep = rProto;
        for (let i = 0; i < 6; i++ ) {
          deep = Object.create(deep);
          Array(20).fill().forEach((_, i) => { deep[`method${i}`] = () => {}; });
        }
      },
      fn: () => { result = serialize(harden(deep)); },
    })
    .on("cycle", evt => console.log(String(evt.target)))
    .on("complete", () => result)
    .run({ async: true })
'
```

</details>

before:
```
shallow remotable x 32,886 ops/sec ±0.96% (89 runs sampled)
deep remotable x 10,695 ops/sec ±1.45% (86 runs sampled)
```

after:
```
shallow remotable x 59,592 ops/sec ±1.61% (84 runs sampled)
deep remotable x 56,173 ops/sec ±2.12% (79 runs sampled)
```