@gibson042 2022-10-21
There are some obstacles to that...
* **module structure**: passStyleOf.js imports from helpers/remotable.js, and IIRC we try to avoid cycles.
* **scope**: the passStyle memo is part of the `makePassStyleOf` closure rather than being a top-level binding.
* **API alignment**: `getInterfaceOf` does not accept a `passStyleOf`/`passStyleOfRecur` argument, and it may be presumptuous to assume use of the one defined in passStyleOf.js.

None of these are insurmountable, but they do give me pause. How important is it to avoid a remotable-specific memoization?