@gibson042 2022-10-23
> Could you file a bug recording your points?

#1339